+++ packages/next-swc/crates/core/Cargo.toml
@@ -21,6 +21,7 @@ regex = "1.5"
 serde = "1"
 serde_json = "1"
 styled_components = {path="../styled_components"}
+modularize_imports = {path="../modularize_imports"}
 swc = "0.161.1"
 swc_atoms = "0.2.11"
 swc_common = { version = "0.17.18", features = ["concurrent", "sourcemap"] }
@@ -31,7 +32,7 @@ swc_node_base = "0.5.1"
 swc_cached = "0.1.1"
 swc_css_prefixer = "0.99.3"
 tracing = {version = "0.1.28", features = ["release_max_level_off"]}
-handlebars = "4.2.1"
+
 
 [dev-dependencies]
 swc_ecma_transforms_testing = "0.75.0"
