+++ packages/next-swc/crates/modularize_imports/src/lib.rs
