+++ packages/next-swc/crates/modularize_imports/tests/fixture/regex/input.js
