+++ packages/next-swc/crates/modularize_imports/tests/fixture/simple/input.js
